+++ lib/src/fitted_text_field_container.dart
@@ -1,9 +1,14 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 
 import 'utils.dart';
 
+typedef double CalculateFunction(TextFieldMeasurer measurer);
+
 class FittedTextFieldContainer extends StatefulWidget {
-  /// Creates a container to wrap a [child] `TextField`  and automatically resizes to fit the [child]'s text value
+  /// Creates an animated container to wrap a [child] `TextField`  and
+  /// automatically resizes to fit the [child]'s text value
   ///
   /// ```dart
   /// FittedTextFieldContainer(
@@ -14,46 +19,41 @@ class FittedTextFieldContainer extends StatefulWidget {
   /// The `TextField` to wrap
   final TextField child;
 
-  /// The width of the `TextField.decoration.prefixIcon` if used
-  final double prefixIconWidth;
+  /// The calculator provides measured values that returns a width value
+  final CalculateFunction calculator;
 
-  /// The width of the `TextField.decoration.suffixIcon` if used
-  final double suffixIconWidth;
-
-  /// The minimum width, default is 0 - which makes the TextField invisible if the TextField
-  /// doesn't include properties that takes up space (such as labelText, prefixIcon, etc).
-  final double minWidth;
-
-  /// The maximum width, defaults to `double.infinity` - i.e. there is no maximum
-  final double maxWidth;
+  /// The builder provides the `child` TextField to a function that returns a
+  /// widget.
+  final Widget Function(BuildContext context, TextField child) builder;
 
   const FittedTextFieldContainer({
     Key key,
     @required this.child,
-    this.prefixIconWidth = 48,
-    this.suffixIconWidth = 48,
-    this.minWidth = 0,
-    this.maxWidth = double.infinity,
+    this.calculator,
+    this.builder,
   })  : assert(child != null),
         super(key: key);
 
   @override
   _FittedTextFieldContainerState createState() =>
       _FittedTextFieldContainerState();
+
+  static CalculateFunction defaultCalculator = (m) =>
+      m.fixedWidths +
+      max(m.labelWidth,
+          m.textField.controller.text.isEmpty ? m.hintWidth : m.textWidth);
 }
 
 class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
-  double _prefixWidth;
-  double _suffixWidth;
-  double _hintWidth;
-  double _labelWidth;
-  double _fixedWidth;
+  CalculateFunction _calculator;
   double _textFieldWidth;
-  TextStyle _defaultTextStyle;
+  TextFieldMeasurer _measurer;
 
   @override
   void initState() {
     assert(widget.child.controller != null);
+    _calculator =
+        widget.calculator ?? FittedTextFieldContainer.defaultCalculator;
     widget.child.controller.addListener(_onTextChanged);
     super.initState();
   }
@@ -62,30 +62,10 @@ class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
   void didChangeDependencies() {
     // When style is null, it defaults to `subtitle1` of current field.
     // See: https://api.flutter.dev/flutter/material/TextField/style.html
-    _defaultTextStyle =
-        widget.child.style ?? Theme.of(context).textTheme.subhead;
+    final textStyle = widget.child.style ?? Theme.of(context).textTheme.subhead;
 
-    _prefixWidth = getPrefixTextSize(widget.child, _defaultTextStyle).width;
-    _suffixWidth = getSuffixTextSize(widget.child, _defaultTextStyle).width;
-    _hintWidth = getHintTextSize(widget.child, _defaultTextStyle).width;
-    _labelWidth = getLabelTextSize(widget.child, _defaultTextStyle).width;
-    _fixedWidth = _prefixWidth + _suffixWidth;
-
-    if (widget.child.decoration.contentPadding != null) {
-      _fixedWidth += widget.child.decoration.contentPadding.collapsedSize.width;
-    }
-
-    if (widget.child.decoration.prefixIcon != null) {
-      _fixedWidth += widget.prefixIconWidth;
-    }
-    if (widget.child.decoration.suffixIcon != null) {
-      _fixedWidth += widget.suffixIconWidth;
-    }
-
-    // Add enough space for the cursor to prevent it being positined onto the next line
-    // in a multiline textfield and scrolled in a single-line text field.
-    _fixedWidth += widget.child.cursorWidth + 1;
-    _textFieldWidth = _geTextFieldWidth();
+    _measurer = TextFieldMeasurer.create(widget.child, textStyle);
+    _textFieldWidth = _calculator(_measurer);
 
     super.didChangeDependencies();
   }
@@ -96,25 +76,8 @@ class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
     super.dispose();
   }
 
-  double _geTextFieldWidth() {
-    double textWidth = getTextSize(widget.child, _defaultTextStyle).width;
-    double width = textWidth > _hintWidth ? textWidth : _hintWidth;
-    if (_labelWidth > width) {
-      width = _labelWidth;
-    }
-    width += _fixedWidth;
-
-    if (width < widget.minWidth) {
-      width = widget.minWidth;
-    }
-    if (width > widget.maxWidth) {
-      width = widget.maxWidth;
-    }
-    return width;
-  }
-
   void _onTextChanged() {
-    final width = _geTextFieldWidth();
+    final width = _calculator(_measurer);
     if (width != _textFieldWidth) {
       setState(() {
         _textFieldWidth = width;
@@ -126,7 +89,9 @@ class _FittedTextFieldContainerState extends State<FittedTextFieldContainer> {
   Widget build(BuildContext context) {
     return Container(
       width: _textFieldWidth,
-      child: widget.child,
+      child: widget.builder == null
+          ? widget.child
+          : widget.builder(context, widget.child),
     );
   }
 }